refactor(faze4): polish po review batch 1 (64-bit byte math, short-read tolerance, dead-code uklid)
readWavRange uziva int64_t offsety + fseeko/_fseeki64 (cross-platform safe
pro WAV >2 GB; streaming worker bude volat s velkymi offsety). Loader uz
nezahazuje sampl pri kratsim WAV (oriznuty soubor: log + pouzi co je).
Smazany dead compat accessory data()/frames_in_ram_head()/total_frames().
Bank.total_frames prejmenovano na Bank.resident_frames (jasnejsi semantika).
TODO k SR-mismatchi v damping crossfadu (latent, faze 6/7). Drobnosti
ze review (#3 invariant komentar, #11 RAII v testu, #12 ciste komentare).

+++ engine/io/wav_reader.h
@@ -6,6 +6,7 @@
 // predlohy z icr (cores/sampler/wav_loader.h), rozdeleny na .h/.cpp a
 // doplneny o peekWavInfo() pro budouci streaming (faze 4).
 
+#include <cstdint>
 #include <string>
 #include <vector>
 
@@ -36,10 +37,11 @@ WavData readWav(const std::string& path);
 WavInfo peekWavInfo(const std::string& path);
 
 // Cte konkretni vyrez WAV souboru → interleaved stereo float. frame_off a
-// frame_count jsou ve stereo frames (ne v samplech). Vrati WavData s
-// frames = min(frame_count, dostupne_do_konce_souboru). Pri offsetu za koncem
+// frame_count jsou ve stereo frames (ne v samplech) — 64-bit, aby streaming
+// worker mohl sahnout do multi-GB WAV souboru s velkymi offsety. Vrati WavData
+// s frames = min(frame_count, dostupne_do_konce_souboru). Pri offsetu za koncem
 // vrati valid=true s frames=0 a prazdnym samples (pro streaming je to OK
 // signal "konec"). Pri chybe otevreni/parsovani vrati valid=false.
-WavData readWavRange(const std::string& path, int frame_off, int frame_count);
+WavData readWavRange(const std::string& path, int64_t frame_off, int64_t frame_count);
 
 } // namespace ithaca
